services/scoring/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
psycopg2-binary==2.9.9
supabase==2.0.2
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
xgboost==2.0.2
tensorflow==2.15.0
python-dotenv==1.0.0
pytest==7.4.3
httpx==0.25.2
openai==1.3.0
celery==5.3.4
redis==5.0.1
reportlab==4.0.7
weasyprint==60.2
sentry-sdk[fastapi]==1.40.0
prometheus-client==0.19.0
prometheus-fastapi-instrumentator==6.1.0